pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "photo-diff"
version = "0.1.0"
description = "CLI POC for image similarity using embedding APIs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "pydantic-settings>=2.3.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
photo-diff = "photo_diff.cli:main"
tile-fetcher = "tile_fetcher.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_diff", "src/tile_fetcher"]
